services/services/pyproject.toml
[tool.poetry]
name = "generic_microservice_services"
version = "0.1.0"
description = "The services submodule for the generic KEVEN microservice, handling gRPC, REST API, and Kafka integration."
authors = ["Craig Holland <craig_holland@hotmail.com>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "generic_microservice_services"}]

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.0.0"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
protobuf = "^4.25.0"
kafka-python = "^2.0.2"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
confluent-kafka = "^2.2.0"

# Internal Dependencies (referencing other submodules)
generic_microservice_domain = "0.1.0"
generic_microservice_database = "0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.1"
mypy = "^1.5.1"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
start = "services.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
